[1912202504] Update hero
To correct the video display on mobile devices, the video has been added in different formats: AVIF, which is like a more advanced GIF; MOV, enabled for Firefox and iOS; and WEBM, the standard format for viewing. Additionally, some extra attributes have been added to address the playback issues on these devices.